pyproject.toml
[project]
name = "ais_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<=3.9.21"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "bleach>=6.2.0",
    "dateparser>=1.2.0",
    "firebase-admin>=6.6.0",
    "flask>=3.1.0",
    "flask-mail>=0.10.0",
    "flask-pydantic>=0.12.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "gevent>=24.11.1",
    "openai>=1.59.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic[email]>=2.10.5",
    "pymysql[rsa]>=1.1.1",
    "pyproject-toml>=0.1.0",
    "selenium>=4.27.1",
    "snowflake-id>=1.0.2",
    "torch>=2.5.1",
    "transformers>=4.47.1",
    "webdriver-manager>=4.0.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]
